Changes based on PR comments
